added GridWar game to test RoundRobinTournament
Grid War is a two-player turn-based zero-sum board game. It is played on
a 4x4 grid. Players take turns to move. Players can move one square at a
time in one of four directions (up, down, left or right). If a player
selects a move that would cause them to exit the board then they forfeit
their turn. A player wins if they move to a square that is already
occupied by their opponent. A player loses if they move in the same
direction on two consecutive moves.

The game is described in the O'Reilly book "Programming Collective
Intelligence" by Tony Segaran (ISBN 978-0-596-52932-1). See:
https://www.safaribooksonline.com/library/view/programming-collective-intelligence/9780596529321/ch11s08.html