Formatting.
Changed tabs to spaces. Rationale (from checkstyle
documentation):

Developers should not need to configure the tab width
of their text editors in order to be able to read source code.
From the
Apache jakarta coding standards: In a distributed development
environment, when the commit messages get sent to a mailing list, they
are almost impossible to read if you use tabs.