renamed Node.getDepth() to getHeight()
Renamed as "get height" is a more accurate description (based on common
terminologies used to describe tree structures) of the purpose of the
method than "get depth" was. The height of a node is the number of edges
on the longest downward path between that node and a leaf. The depth of
a node is the number of edges from the node to the tree's root node.
Source: http://en.wikipedia.org/wiki/Tree_%28data_structure%29